1、更新http工具 2、更新字符串处理工具

+++ types.h
@@ -16,11 +16,11 @@ typedef unsigned char bool;
 /* define int types*/
 typedef     signed char         int8_t;
 typedef     signed short        int16_t;
-typedef     signed int          int32_t;
+typedef     signed long			int32_t;
 typedef     signed __int64      int64_t;
 typedef     unsigned char       uint8_t;
 typedef     unsigned short      uint16_t;
-typedef     unsigned int        uint32_t;
+typedef     unsigned long       uint32_t;
 typedef     unsigned __int64    uint64_t;
 
 /* the following definitions are from VS2010's stdint.h */
